+++ packages/lambda/src/functions/helpers/validate-composition.ts
@@ -4,7 +4,7 @@ import type {
 	openBrowser,
 } from '@remotion/renderer';
 import {getCompositions} from '@remotion/renderer';
-import type {TCompMetadata} from 'remotion';
+import type {DownloadMap, TCompMetadata} from 'remotion';
 import type {Await} from '../../shared/await';
 
 type ValidateCompositionOptions = {
@@ -18,6 +18,7 @@ type ValidateCompositionOptions = {
 	timeoutInMilliseconds: number;
 	chromiumOptions: ChromiumOptions;
 	port: number | null;
+	downloadMap: DownloadMap;
 };
 
 export const validateComposition = async ({
@@ -31,6 +32,7 @@ export const validateComposition = async ({
 	ffprobeExecutable,
 	chromiumOptions,
 	port,
+	downloadMap,
 }: ValidateCompositionOptions): Promise<TCompMetadata> => {
 	const compositions = await getCompositions(serveUrl, {
 		puppeteerInstance: browserInstance,
@@ -41,6 +43,7 @@ export const validateComposition = async ({
 		timeoutInMilliseconds,
 		chromiumOptions,
 		port,
+		downloadMap,
 	});
 
 	const found = compositions.find((c) => c.id === composition);
