+++ packages/renderer/src/render-media.ts
@@ -2,9 +2,10 @@ import type {ExecaChildProcess} from 'execa';
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
-import type {SmallTCompMetadata} from 'remotion';
+import type {DownloadMap, SmallTCompMetadata} from 'remotion';
 import {Internals} from 'remotion';
 import type {RenderMediaOnDownload} from './assets/download-and-map-assets-to-file';
+import {makeDownloadMap} from './assets/download-map';
 import type {BrowserExecutable} from './browser-executable';
 import type {BrowserLog} from './browser-log';
 import type {Browser as PuppeteerBrowser} from './browser/Browser';
@@ -79,6 +80,10 @@ export type RenderMediaOptions = {
 	cancelSignal?: CancelSignal;
 	browserExecutable?: BrowserExecutable;
 	verbose?: boolean;
+	/**
+	 * @deprecated Only for Remotion internal usage
+	 */
+	downloadMap?: DownloadMap;
 } & ServeUrlOrWebpackBundle;
 
 type Await<T> = T extends PromiseLike<infer U> ? U : T;
@@ -116,6 +121,7 @@ export const renderMedia = ({
 	browserExecutable,
 	port,
 	cancelSignal,
+	downloadMap,
 	...options
 }: RenderMediaOptions): Promise<Buffer | null> => {
 	validateQuality(quality);
@@ -280,6 +286,7 @@ export const renderMedia = ({
 				browserExecutable,
 				port,
 				cancelSignal: cancelRenderFrames.cancelSignal,
+				downloadMap: downloadMap ?? makeDownloadMap(),
 			});
 
 			return renderFramesProc;
