+++ packages/renderer/src/stitch-frames-to-video.ts
@@ -2,12 +2,13 @@ import execa from 'execa';
 import fs, {unlinkSync} from 'fs';
 import {readFile} from 'fs/promises';
 import path from 'path';
-import type {DownloadMap, RenderAssetInfo, TAsset} from 'remotion';
+import type {DownloadMap, TAsset} from 'remotion';
 import {Internals} from 'remotion';
 import {calculateAssetPositions} from './assets/calculate-asset-positions';
 import {convertAssetsToFileUrls} from './assets/convert-assets-to-file-urls';
 import type {RenderMediaOnDownload} from './assets/download-and-map-assets-to-file';
 import {markAllAssetsAsDownloaded} from './assets/download-and-map-assets-to-file';
+import type {RenderAssetInfo} from './assets/download-map';
 import type {Assets} from './assets/types';
 import type {Codec} from './codec';
 import {DEFAULT_CODEC} from './codec';
