+++ packages/renderer/src/test/render-still.test.ts
@@ -1,4 +1,5 @@
 import {expect, test} from 'vitest';
+import {makeDownloadMap} from '../assets/download-map';
 import {renderStill} from '../render-still';
 
 test('Need to pass valid metadata', () => {
@@ -15,6 +16,7 @@ test('Need to pass valid metadata', () => {
 			frame: 0,
 			output: '/file/output.png',
 			serveUrl: 'https://silly-crostata-c4c336.netlify.app/',
+			downloadMap: makeDownloadMap(),
 		})
 	).rejects.toThrow(/not be NaN, but is NaN/);
 });
@@ -33,6 +35,7 @@ test('Need to pass valid metadata', () => {
 			frame: 200,
 			output: '/file/output.png',
 			serveUrl: 'https://silly-crostata-c4c336.netlify.app/',
+			downloadMap: makeDownloadMap(),
 		})
 	).rejects.toThrow(
 		/Cannot use frame 200: Duration of composition is 30, therefore the highest frame that can be rendered is 29/
